Quick Toggle : Modular Avatar version
Creating a toggle that you can put in a menu, in a rather
flexible way is a FUCKING PAIN WITH THIS SYSTEM,
WHAT THE HELL HAPPENED DURING THE CREATION OF THIS THING !?

Like, seriously, What the hell ! It took me several attempts
to understand how fucked up Modular Avatar can be when it comes
to Menu creation and handling basic setup...

Ugh...

Still, this 'works' and handles menus efficiently.
Also everything that's generated can be used elsewhere.

Signed-off-by: Voyage <voyage@miouyouyou.fr>